Changes to Service and Repo files
I added all required method signatures to remove red errors in project. I commented out anything that had to do with likes because I don't know how those will be implemented. The tests do not run properly due to a missing or wrong configuration within the Test class, but the tests are logically what they're supposed to be.

I do not know if the error creating the User table is having an effect on the Post classes but it may be considered.

As of right now we can check and fix the logic for the methods I wrote in Service and Repository because I'm sure some are wrong.

If the Service logic is correct, we can start on the Controller as well.

See if you can figure out the configuration for the Post JUnit tests. I did not write regular JUnit tests in Spring during the last project so I'm not sure what annotations I needed.

We still need JUnit test classes for Comments and PostFeed. I don't know how those two will work or be organized so I have not planned anything for them as of yet.